Apply to inspection warnings;

+++ Ship/Sailor.cpp
@@ -3,12 +3,8 @@
 #include "Sailor.h"
 #include "Mast.h"
 #include "Ship.h"
-#include "../Wind.h"
-#include "../World.h"
 #include "MastDistributor.h"
-#include "../Util/Util.h"
 #include "Cannon.h"
-#include "ShipLayout.h"
 
 using namespace std;
 
@@ -191,7 +187,7 @@ Vec2f Sailor::CalculateCannonTarget() const {
     float distance = 5;
     if(cannon_target != nullptr)
         distance = (cannon_target->GetPosition() - parent->GetPosition()).Length();
-    Vec2f perpendicular_right = Vec2f::FromAngle(parent->GetDirection().Angle() + M_PI_2).Normalized() * distance;
+    Vec2f perpendicular_right = Vec2f::FromAngle(parent->GetDirection().Angle() + (float)M_PI_2).Normalized() * distance;
     return parent->GetPosition() + perpendicular_right * (use_right_cannons ? 1.f : -1.f);
 }
